+++ multivar_horner/multivar_horner.py
@@ -1,14 +1,17 @@
-# TODO
-# faster tree build
-# explain parameters flags
-# matlab binding
-# TODO test gradient
-# TODO multivariate newton raphson method
-# TODO mention in readme
-# TODO function to predict the factorisation time based on dim, max_degree and num_entries
-# TODO based on system
-# TODO suggest function to use
-# TODO MATH: find algorithm to parse optimal tree (no procedure known for this!)
+# TODO s:
+
+# matlab binding?
+# optimize naive version without factorisation: compute all needed exponents only once. order exponents cleverly.
+# store at which index the later entries (exponents in column) will just be 0 -> reduce the matrix size piecewise
+
+# test gradient
+# multivariate newton raphson method
+# mention new features in readme
+# function to predict the factorisation time and efficiency (time saved) based on dim, max_degree and num_entries
+# based on system
+# suggest function to use
+
+# test factorisation for univariate polynomials! should always find optimum (unique 1D horner fact.)
 
 
 import itertools
@@ -16,15 +19,16 @@ import pickle
 
 import numpy as np
 
+from .factorisation_classes import HeuristicFactorisationRoot, OptimalFactorisationRoot
 from .global_settings import DEBUG, DEFAULT_PICKLE_FILE_NAME, FLOAT_DTYPE, UINT_DTYPE
-from .helper_classes import HornerTree, ScalarFactor
+from .helper_classes import FactorContainer, ScalarFactor
 from .helper_fcts import get_prime_array, rectify, validate
-from .helpers_fcts_numba import eval_naive, eval_recipe
+from .helpers_fcts_numba import eval_recipe, naive_eval
 
 
 # is not a helper function to make it an importable part of the package
 def load_pickle(path=DEFAULT_PICKLE_FILE_NAME):
-    print('importing polygon from file "{}"'.format(path))
+    print('importing polynomial from file "{}"'.format(path))
     with open(path, 'rb') as f:
         return pickle.load(f)
 
@@ -41,7 +45,8 @@ class MultivarPolynomial(object):
     """
 
     # prevent dynamic attribute assignment (-> safe memory)
-    __slots__ = ['coefficients', 'exponents', 'dim', 'order', 'max_degree', 'unused_variables', 'representation']
+    __slots__ = ['coefficients', 'exponents', 'existing_exponents', 'num_monomials', 'dim', 'order', 'max_degree',
+                 'unused_variables', 'representation', 'heuristic']
 
     def __init__(self, coefficients, exponents, rectify_input=False, validate_input=False):
         """
@@ -61,12 +66,15 @@ class MultivarPolynomial(object):
         self.coefficients = coefficients
         self.exponents = exponents
 
+        self.num_monomials = self.exponents.shape[0]
         self.dim = self.exponents.shape[1]
         self.order = np.sum(self.exponents, axis=0).max()
         self.max_degree = self.exponents.max()
 
         self.unused_variables = np.where(~np.any(self.exponents, axis=1))[0]
         self.representation = ''
+        self.existing_exponents = None
+        self.heuristic = None
 
     def __str__(self):
         return self.get_representation()
@@ -75,13 +83,15 @@ class MultivarPolynomial(object):
         return self.__str__()
 
     def get_representation(self):
-        s = '[{}] p(x) = '.format(self.get_num_ops())
+
+        s = '[#ops={}] p(x) = '.format(self.get_num_ops())
         monomials = []
         for i, exp_vect in enumerate(self.exponents):
             monomial = [str(self.coefficients[i, 0])]
             for dim, exp in enumerate(exp_vect):
-                if exp > 0:
-                    monomial.append('x_{}^{}'.format(dim + 1, exp))
+                # if exp > 0:
+                # show all operations, even 1 * x_i^0
+                monomial.append('x_{}^{}'.format(dim + 1, exp))
             monomials.append(' '.join(monomial))
 
         s += ' + '.join(monomials)
@@ -113,21 +123,24 @@ class MultivarPolynomial(object):
             assert len(x.shape) == 1
             assert x.shape[0] == self.dim
 
-        return eval_naive(x, self.coefficients, self.exponents)
+        return naive_eval(x, self.coefficients, self.exponents)
 
-    def partial_derivative(self, i):
+    def get_partial_derivative(self, i):
         """
         TODO test
         f(x) = a x^b
         f'(x) = ab x^(b-1)
         :param i:
         :return: the partial derivative of this polynomial wrt. the i-th coordinate
+        # ATTENTION: coordinate counting starts with 1! (first coordinates is #1)
         """
 
-        # IMPORTANT: do not modify the stored coefficient and exponent arrays!
+        assert (0 < i <= self.dim)
+        coord_index = i - 1
+        # IMPORTANT: do not modify the stored coefficient and exponent arrays of self!
         # set all the coefficients not depending on the i-th coordinate to 0
-        # this simply means not adding them to the list of coefficients
-        active_rows = np.where(self.exponents[:, i] >= 1)[0]
+        # this simply means not adding them to the list of coefficients of the new polynomial class instance
+        active_rows = np.where(self.exponents[:, coord_index] >= 1)[0]
 
         new_coefficients = self.coefficients[active_rows]
         new_exponents = self.exponents[active_rows, :]
@@ -138,18 +151,21 @@ class MultivarPolynomial(object):
             assert new_exponents.shape[1] == self.dim and len(new_exponents.shape) == 2
 
         # multiply the coefficients with the exponent of the i-th coordinate
-        new_coefficients *= new_exponents[:, i]
+        new_coefficients = np.multiply(new_coefficients.flatten(), new_exponents[:, coord_index])
+        new_coefficients = new_coefficients.reshape(-1, 1)
 
         # reduce the the exponent of the i-th coordinate by 1
-        new_exponents[:, i] -= 1
+        new_exponents[:, coord_index] -= 1
 
+        # must call the proper constructor method also for inherited classes
+        # TODO also use option 'univariate_factors'
         return self.__class__(new_coefficients, new_exponents)
 
-    def gradient(self):
+    def get_gradient(self):
         """
         :return: the list of all partial derivatives
         """
-        return [self.partial_derivative(i) for i in range(self.dim)]
+        return [self.get_partial_derivative(i) for i in range(1, self.dim + 1)]
 
 
 class HornerMultivarPolynomial(MultivarPolynomial):
@@ -164,12 +180,11 @@ class HornerMultivarPolynomial(MultivarPolynomial):
     """
     # __slots__ declared in parents are available in child classes. However, child subclasses will get a __dict__
     # and __weakref__ unless they also define __slots__ (which should only contain names of any additional slots).
-    __slots__ = ['prime_array', 'horner_tree', 'unique_factor_id_list', 'unique_factors', 'num_ops', 'subtree_amount',
-                 'initial_value_array', 'copy_recipe', 'scalar_recipe', 'monomial_recipe', 'tree_recipe',
-                 'tree_ops']
+    __slots__ = ['prime_array', 'factorisation_tree', 'factor_container', 'num_ops', 'initial_value_array',
+                 'copy_recipe', 'scalar_recipe', 'monomial_recipe', 'tree_recipe', 'tree_ops']
 
     def __init__(self, coefficients, exponents, rectify_input=False, validate_input=False, keep_tree=False,
-                 univariate_factors=False):
+                 compute_representation=False, find_optimal=False):
         """
         :param coefficients: a numpy array column vector of doubles representing the coefficients of the monomials
         :param exponents: a numpy array matrix of unsigned integers representing the exponents of the monomials
@@ -179,21 +194,17 @@ class HornerMultivarPolynomial(MultivarPolynomial):
         """
 
         def compute_num_ops():
-            # count the number of instructions done when computing all factors
-            # ... and when evaluating the horner factorisation
-
-            # do NOT recursively compute for every subtree
-            # num_ops = 0
-            # for f in self.unique_factors:
-            #     num_ops += f.num_ops()
-            #
-            # num_ops += self.horner_tree.num_ops()
+            # count the number of instructions done during evaluation
+            # (computing all factors + evaluating the factorisation subtrees)
             # for every entry in each recipe one operation is being done
-            self.num_ops = self.copy_recipe.shape[0] + self.scalar_recipe.shape[0] + self.monomial_recipe.shape[0] + \
-                           self.tree_ops.shape[0]
+            # do not count "copy" instructions (not really a mathematical operation)
+            self.num_ops = self.scalar_recipe.shape[0] + self.monomial_recipe.shape[0] + self.tree_ops.shape[0]
 
         def get_string_representation():
-            return '[{}] p(x) = '.format(self.num_ops) + self.horner_tree.__str__()
+            representation = '[#ops={}] p(x)'.format(self.num_ops)
+            if compute_representation:
+                representation += ' = ' + self.factorisation_tree.__str__()
+            return representation
 
         super(HornerMultivarPolynomial, self).__init__(coefficients, exponents, rectify_input, validate_input)
 
@@ -201,18 +212,18 @@ class HornerMultivarPolynomial(MultivarPolynomial):
         self.prime_array = get_prime_array(self.dim)
 
         # store all unique factors of the horner factorisation
-        # NOTE: do NOT create all scalar factors with exponent 1 (they might unused!)
-        self.unique_factor_id_list = []
-        self.unique_factors = []
+        # NOTE: do NOT create all scalar factors with exponent 1 (they might be unused!)
+        self.factor_container = FactorContainer(self.prime_array)
 
         # factorize the polynomial once and store the factorisation as a tree
-        subtree_id_cntr = itertools.count(0)
+        id_counter = itertools.count(0)
         tree_coefficients = []
-        self.horner_tree = HornerTree(self.coefficients, self.exponents, self.prime_array,
-                                      self.unique_factor_id_list, self.unique_factors, tree_coefficients,
-                                      id_counter=subtree_id_cntr, univariate_factors=univariate_factors)
-        # the all subtrees now got their id. the next value from the counter is equal to the subtree amount
-        self.subtree_amount = subtree_id_cntr.__next__()
+        if find_optimal:
+            self.factorisation_tree = OptimalFactorisationRoot(self.coefficients, tree_coefficients, id_counter,
+                                                               self.factor_container, exponents=self.exponents)
+        else:
+            self.factorisation_tree = HeuristicFactorisationRoot(self.coefficients, tree_coefficients, id_counter,
+                                                                 self.factor_container, exponents=self.exponents)
 
         # factor list is now filled with the unique factors
         # during evaluation of a polynomial the values of all the unique factors are needed at least once
@@ -224,18 +235,17 @@ class HornerMultivarPolynomial(MultivarPolynomial):
 
         # compile and store a "recipe" for evaluating the polynomial with just numpy arrays
         self.initial_value_array, self.copy_recipe, self.scalar_recipe, self.monomial_recipe, self.tree_recipe, \
-        self.tree_ops = self.compile_recipes(tree_coefficients)
+            self.tree_ops = self.compile_recipes(tree_coefficients)
 
         compute_num_ops()
         self.representation = get_string_representation()
 
         if not keep_tree:
-            # the trees and factors are not being needed any more
+            # the trees and factors are not needed any more
             # a value lookup can be done with just the recipe
             # free up the memory
-            del self.horner_tree
-            del self.unique_factors
-            del self.unique_factor_id_list
+            del self.factorisation_tree
+            del self.factor_container
             del self.prime_array
 
     def __str__(self):
@@ -246,7 +256,6 @@ class HornerMultivarPolynomial(MultivarPolynomial):
 
     def link_monomials(self):
         """
-        TODO numba precompile
         find the optimal factorisation of the unique factors themselves
         since the monomial ids are products of the ids of their scalar factors
         check for the highest possible divisor among all factor ids
@@ -254,25 +263,26 @@ class HornerMultivarPolynomial(MultivarPolynomial):
         :return:
         """
         # sort after their id
-        self.unique_factor_id_list = list(sorted(self.unique_factor_id_list))
-        self.unique_factors = list(sorted(self.unique_factors, key=lambda f: f.monomial_id))
+        unique_factor_id_list = list(sorted(self.factor_container.id2idx.keys()))
+        unique_factors = list(sorted(self.factor_container.factors, key=lambda f: f.monomial_id))
         # property of the list: the scalar factors of each monomial are stored in front of it
 
         # IMPORTANT: properly set the indices of the values for each factor
-        # the values of every factor are stored after the coefficients of all subtrees
-        value_idx_offset = self.subtree_amount
-        for idx, f in enumerate(self.unique_factors):
-            f.value_idx = idx + value_idx_offset
+        # the values of every factor are stored after the coefficients
+        value_idx_offset = self.num_monomials
+        for idx, factor in enumerate(unique_factors):
+            factor.value_idx = idx + value_idx_offset
 
         # start at the last factor (highest id)
-        pointer1 = len(self.unique_factors) - 1
+        pointer1 = len(unique_factors) - 1
 
         # the smallest factor has no factorisation (stop at pointer=1)
         while pointer1 > 0:
-            candidate = self.unique_factors[pointer1]
+            candidate = unique_factors[pointer1]
 
             if type(candidate) is ScalarFactor:
-                # scalar factors have no factorisation
+                # scalar factors have no useful factorisation
+                # (can be directly computed with the same amount of operations)
                 pointer1 -= 1
                 continue
 
@@ -291,12 +301,12 @@ class HornerMultivarPolynomial(MultivarPolynomial):
                     # picking a factorizing monomial cannot be factorised itself
                     # just pick the scalar factors of the monomial
                     # ATTENTION: offset needed!
-                    factorisation_idxs = [self.unique_factor_id_list.index(scalar_factor.monomial_id) + value_idx_offset
+                    factorisation_idxs = [unique_factor_id_list.index(scalar_factor.monomial_id) + value_idx_offset
                                           for scalar_factor
                                           in candidate.scalar_factors]
                     break
 
-                monomial_id2 = self.unique_factors[pointer2].monomial_id
+                monomial_id2 = unique_factors[pointer2].monomial_id
 
                 if remaining_factor_id < monomial_id2:
                     # this monomial has a higher id than the remaining factor. it cannot be a factorisation
@@ -334,14 +344,12 @@ class HornerMultivarPolynomial(MultivarPolynomial):
 
         # for evaluating the polynomial in tree form intermediary computation results have to be stored in a value array
         # the value array has one entry for every subtree and one for every factor
-        value_array_length = self.subtree_amount + len(self.unique_factors)
+        value_array_length = self.num_monomials + len(self.factor_container.factors)
 
         initial_value_array = np.empty(value_array_length, dtype=FLOAT_DTYPE)
 
         # initial_value_array = np.array(tree_coefficients, dtype=FLOAT_DTYPE)..reshape(value_array_length)
         # the initial value array has the coefficients of all subtrees stored at the index of their id
-        # TODO
-        # initial_value_array[0:len(tree_coefficients)-1] = tree_coefficients
         for i, v in enumerate(tree_coefficients):
             initial_value_array[i] = v
 
@@ -353,25 +361,24 @@ class HornerMultivarPolynomial(MultivarPolynomial):
         copy_recipe = []  # skip computing factors with exp 1, just copy x vals
         scalar_recipe = []
         monomial_recipe = []
-        for f in self.unique_factors:
-            factor_recipe = f.get_recipe()
-            if type(f) is ScalarFactor:
-                copy_recipe += factor_recipe[0]
-                scalar_recipe += factor_recipe[1]
-            else:
-                monomial_recipe += factor_recipe
+        for f in self.factor_container.factors:
+            copy_instr, scalar_instr, monomial_instr = f.get_recipe()
+            copy_recipe += copy_instr
+            scalar_recipe += scalar_instr
+            monomial_recipe += monomial_instr
 
         # compile the recipe for evaluating the horner factorisation tree
-        tree_recipe, tree_ops = self.horner_tree.get_recipe()
+        tree_recipe, tree_ops = self.factorisation_tree.get_recipe()
         # convert and store the recipes
         # for the recipes numba is expecting the data types:
         #   array(uint, 2d, C),
         #   separate boolean array for operations, uint not needed (just 0 or 1)
-        return initial_value_array, np.array(
-            copy_recipe, dtype=UINT_DTYPE).reshape((-1, 2)), np.array(
-            scalar_recipe, dtype=UINT_DTYPE).reshape((-1, 3)), np.array(
-            monomial_recipe, dtype=UINT_DTYPE).reshape((-1, 3)), np.array(
-            tree_recipe, dtype=UINT_DTYPE).reshape((-1, 2)), np.array(tree_ops, dtype=np.bool)
+        return (initial_value_array,
+                np.array(copy_recipe, dtype=UINT_DTYPE).reshape((-1, 2)),
+                np.array(scalar_recipe, dtype=UINT_DTYPE).reshape((-1, 3)),
+                np.array(monomial_recipe, dtype=UINT_DTYPE).reshape((-1, 3)),
+                np.array(tree_recipe, dtype=UINT_DTYPE).reshape((-1, 2)),
+                np.array(tree_ops, dtype=np.bool))
 
     def eval(self, x, validate_input=False):
         """